test coherence avec opt-2.7b On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   open_text_gen/compute_coherence.py
	new file:   open_text_gen/wikitext_grid_search/wikitext_k5_a0.6_e0.0_gpt2._opt-2.7b_coherence_result.json

+++ open_text_gen/compute_coherence.py
@@ -55,7 +55,7 @@ class CoherenceEvaluator(nn.Module):
         outputs = self.model(input_ids=input_ids, output_hidden_states=True)
         logits = outputs.logits
         assert logits.size() == torch.Size([bsz, seqlen, self.vocab_size])
-        probability = F.softmax(logits, dim=-1) # bsz x seqlen x vocab_size
+        probability = F.softmax(logits.float(), dim=-1) # bsz x seqlen x vocab_size
         last_hidden_states = outputs.hidden_states[-1]
         assert last_hidden_states.size() == torch.Size([bsz, seqlen, self.embed_dim])
         return last_hidden_states, probability
